+++ rabbitmq_class.py
@@ -72,12 +72,15 @@ def create_rmqcon(cfg, q_name, r_key):
 
     """
 
+    heartbeat = cfg.heartbeat if hasattr(cfg, "heartbeat") else 60
+    host_list = cfg.host_list if hasattr(cfg, "host_list") else list()
+
     return RabbitMQCon(
         cfg.user, cfg.japd, cfg.host, cfg.port,
         exchange_name=cfg.exchange_name, exchange_type=cfg.exchange_type,
         queue_name=q_name, routing_key=r_key, x_durable=cfg.x_durable,
         q_durable=cfg.q_durable, auto_delete=cfg.auto_delete,
-        no_ack=cfg.no_ack)
+        no_ack=cfg.no_ack, heartbeat=heartbeat, host_list=host_list)
 
 
 def create_rmqpub(cfg, q_name, r_key):
@@ -344,14 +347,28 @@ class RabbitMQPub(RabbitMQ):
         Arguments:
             (input) body -> Message body being published to RabbitMQ.
             (input) mandatory -> True|False - Message is saved to queue.
-            (output) True|False -> Message confirmation delivery status.
+            (output) status -> True|False -> Message confirmation delivery.
 
         """
 
-        return self.channel.basic_publish(
-            exchange=self.exchange, routing_key=self.routing_key,
-            body=body, mandatory=mandatory,
-            properties=pika.BasicProperties(delivery_mode=2))
+        if pika.__version__ < '1.0.0':
+            status = self.channel.basic_publish(
+                exchange=self.exchange, routing_key=self.routing_key,
+                body=body, mandatory=mandatory,
+                properties=pika.BasicProperties(delivery_mode=2))
+
+        else:
+            try:
+                self.channel.basic_publish(
+                    exchange=self.exchange, routing_key=self.routing_key,
+                    body=body, mandatory=mandatory,
+                    properties=pika.BasicProperties(delivery_mode=2))
+                status = True
+
+            except pika.exceptions.UnroutableError:
+                status = False
+
+        return status
 
     def setup_queue(self):