+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.1.1] - 2021-09-29
+### Fixed
+- RabbitMQPub.publish_msg:  Added version check, due to change in Pika's basic_publish return status.
+- create_rmqcon, create_rmqpub:  Added heartbeat and host_list to RabbitMQ instance call.
+
+
 ## [2.1.0] - 2021-09-28
 - Update to work with Pika 1.2.0
 - Update to work with RabbitMQ 3.8.2